By speeding up a handful of the longest expressions to compile.

Here are some quick unscientific measurements reported by `make display_compilation_time`:

| Declaration                                                    | Before   | After    | Speedup |
| -------------------------------------------------------------- | -------- | -------- | ------- |
| `NimbleOperatorRule.swift/violationMatchesRanges(in:)`         | 244.53ms | 5.08ms   | 48x     |
| `LintableFilesVisitor.swift/resolveParamsFiles(args:)`         | 251.44ms | 106.95ms | 2.4x    |
| `TestHelpers.swift/render(violations:in:)`                     | 295.11ms | 125.02ms | 2.4x    |
| `FunctionBodyLengthRule.swift/validate(file:kind:dictionary:)` | 312.87ms | 186.87ms | 1.7x    |
| `InertDeferRule.swift/validate(file:)`                         | 315.05ms | 198.54ms | 1.6x    |